+++ mach1/mach1.f90
@@ -6,8 +6,8 @@ PROGRAM mach1
     INTEGER :: n, localn
     INTEGER*8 :: i
     CHARACTER(32) :: argv
-    REAL, dimension(:), allocatable :: vector, localvector, globalvector
-    REAL :: frac1, frac2, acfrac1, acfrac2, tiny
+    DOUBLE PRECISION, dimension(:), allocatable :: vector, localvector, globalvector
+    DOUBLE PRECISION :: frac1, frac2, acfrac1, acfrac2, tiny
     LOGICAL :: localequal, equal 
 
     tiny = 1e-16
